+++ src/Controller/BoxAdLoginController.php
@@ -91,7 +91,7 @@ class BoxAdLoginController extends AbstractController
             return new RedirectResponse($boxUrl);
         } catch (\Exception|InvalidArgumentException $exception) {
             $error = $request->query->get('error');
-            $errorDescription = urldecode($request->query->get('error_description'));
+            $errorDescription = urldecode($request->query->get('error_description', ''));
 
             $session->set('exceptionMessage', $exception->getMessage());
             $session->set('error', 'Azure Ad: '.$error);