+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added calls to getToken when missing or invalid token.
 - Added reload browser when connection has been shut down by the server.
+- Fixed urldecode of null error in oidc error handling.
 
 ## [1.3.0] - 2024-11-07